Refactor: moved NumbersConditions and its tests
